fix(review): strtol int64_t cast; version bump in README+SRS; changelog REST note
MISRA-1.2: parser_meta.cpp — change int32_t to int64_t in all three
strtol-based parse functions (parsePulseMinAltLineLocked,
parsePulseSafeDelayLineLocked, parsePulseArmTimeoutLineLocked).
int32_t cast before range check allowed wrap-around on 64-bit hosts
(e.g. 4294992296 -> 25000 would pass [1,50000] check); int64_t holds
the full strtol return value on all supported platforms.
Cast to uint32_t only after bounds check passes.

docs: bump version references from 2.2.5 to 2.2.6 in README.md badge
and docs/requirements/SRS.sdoc (TITLE, VERSION, body reference).

docs: update v2.2.6 changelog — type and summary now call out the
HTTP 503->500 behavior change on /api/pulse/status; MISRA-1.2 entry
explains the int64_t rationale with a concrete wrap-attack example.

+++ src/ams/mission_script_engine_parser_meta.cpp
@@ -486,7 +486,7 @@ bool MissionScriptEngine::parsePulseMinAltLineLocked(const char* line)
     if (*rest == '\0') { setErrorLocked("pulse.min_altitude: missing value"); return false; }
 
     char* endp = nullptr;
-    const int32_t val = static_cast<int32_t>(strtol(rest, &endp, 10));
+    const int64_t val = static_cast<int64_t>(strtol(rest, &endp, 10));
     if (endp == rest || val <= 0 || val > 50000)
     {
         setErrorLocked("pulse.min_altitude: value must be in range [1, 50000] m");
@@ -526,7 +526,7 @@ bool MissionScriptEngine::parsePulseSafeDelayLineLocked(const char* line)
     if (*rest == '\0') { setErrorLocked("pulse.safe_delay: missing value"); return false; }
 
     char* endp = nullptr;
-    const int32_t val = static_cast<int32_t>(strtol(rest, &endp, 10));
+    const int64_t val = static_cast<int64_t>(strtol(rest, &endp, 10));
     if (endp == rest || val < 100 || val > 60000)
     {
         setErrorLocked("pulse.safe_delay: value must be in range [100, 60000] ms");
@@ -565,7 +565,7 @@ bool MissionScriptEngine::parsePulseArmTimeoutLineLocked(const char* line)
     if (*rest == '\0') { setErrorLocked("pulse.arm_timeout: missing value"); return false; }
 
     char* endp = nullptr;
-    const int32_t val = static_cast<int32_t>(strtol(rest, &endp, 10));
+    const int64_t val = static_cast<int64_t>(strtol(rest, &endp, 10));
     if (endp == rest || val < 500 || val > 300000)
     {
         setErrorLocked("pulse.arm_timeout: value must be in range [500, 300000] ms");
